fix(datetimepickerv2): revert changes in timepickerdropdown
